Fix issue where geocode distance was broken on ES >= 1.4.
After upgrading to ElasticSearch 1.4, or later, scripting is disabled by
default. Furthermore, the way dynamic scripting is specified for it to
be returned as part of the resultset also changed. This commit adds
support for Groovy scripting to be used, as it is the recommended way
for ES scripting (since it is sandboxed.) It also maintains backwards
compatibility for ES versions prior to 1.4.

No change is needed on the model for this to work, but a change is
needed in the ES server settings for this to work. This is not
considered a show-stopper since on those servers the previous dynamic
fields were not working anyway. To enable groovy scripting support, the
following settings are recommended (normally to be added on the
`elasticsearch.yml` configuration file):

```yaml
script.disable_dynamic: "sandbox"
script.groovy.sandbox.enabled: true
```

This adds support for dynamic scripting **ONLY** with languages that
support sandboxing, and enables sandbox on Groovy.